services/web_master/requirements.txt
fastapi
uvicorn[standard]
pydantic
pyyaml
requests
pandas
scikit-learn
joblib
